:art: パースエラーの定義を lib.rs から分離した
